Early milestone submission to be verified by project team members
Session system is completely working and session validation is done on all get and post routes so far. User authentication has been implemented at the login post route to determine if a user is an admin or member, and alter the role in the session data accordingly. The member page or admin page is served to the user based on his role. Flash messages have also been implemented to display customer confirmation and error messages to the user depending  on the currently rendered page. Moreover, login form validation has been done to validate that the user does not leave any empty fields. CSRF protection functions have been created in a separate file, but has not been implemented for any forms in any route so far. Additionally, the core-ui template has been included for all rendered web pages, and the list of feeding sites are dynamically rendered from the database when any public viewer accesses the website.