Added a 404 page to handle non-existent routes + Session handling for GET "/" route
A 404 not found page is rendered each time the user tries to access a route which does not exist on the server. Also, session handling has been implemented for the GET "/" route to handle the case where the user does not have a browser cookie or valid user session in the database.