Added log out functionality + user login and logout alert modal + "Back to landing page" button to login and register pages + Updated email validation + Custom nav bar for public viewer, member, and admin
Members and admins now have a logout button available in their nav bar. This allows them to log out of their account which terminates their current session and redirects them to the landing page, where a new public viewer session is created for them.

Moreover, an alert modal has been created and is displayed every time a user successfully logs in to a new account, greeting him/her with a custom message containing his/her username. Another alert modal is also generated when a user logs out  of his account, thanking the user for his visit to the site.

In addition to this, a button which redirects a user to the landing page (similar to the one found in the 404 page) has been added in both the login and registration pages. This was done to allow users to go back to the landing page without using the back button in the top left corner of the  browser tab.

Furthermore, the regular expression string added by my team member @Khalid-60100381  has been slightly tweaked to force the user to only enter a single @ symbol, separating the local-part and the domain. Additionally, only a single dot has been allowed which distinguishes the domain and top-level domain in the email.

Finally, a public viewer, member, and admin will each have a custom navigation bar as they move from page to page. This was done to separate concerns between the three roles and only provide navigation links that are relevant to the user's role. Having the same navigation bar stick with the user as he/she navigates the website was also a quality of life decision which allows the user to move from one page to another without having to go back to one centralized page each time which contains all the navigation links.