Updated a bug where the changed password was not being correctly hashed in the database
There was an issue where the user's modified password in the change profile details page was being persisted as plaintext in the database. Managed to fix this by creating a regex which validates if the user did not change his password by checking that the user's current password is a SHA-512 hash and is not plaintext, else it hashes the user's modified password.